[COLAB-2918]: remove @EnableFeignClients from UserDaoTest; Move @EnableFeignClients to separate class to be able to distinguish non-test and test configurations; Test-config: TrackingClientMockImpl
