+++ microservices/services/members-service/src/main/java/org/xcolab/service/members/FeignEnable.java
@@ -1,0 +1,12 @@
+package org.xcolab.service.members;
+
+import org.springframework.cloud.netflix.feign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+
+@Configuration
+@EnableFeignClients(basePackages = { "org.xcolab.client.tracking" })
+@Profile("!test")
+public class FeignEnable {
+
+}
