+++ microservices/services/members-service/src/main/java/org/xcolab/service/members/MembersServiceApplication.java
@@ -8,7 +8,6 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.netflix.feign.EnableFeignClients;
 
 @SpringBootApplication
-@EnableFeignClients(basePackages = { "org.xcolab.client.tracking" })
 @EnableDiscoveryClient
 public class MembersServiceApplication extends SpringBootServletInitializer {
