+++ view/src/main/java/org/xcolab/view/ColabViewApplication.java
@@ -8,7 +8,6 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.netflix.feign.EnableFeignClients;
 
 @SpringBootApplication
-@EnableFeignClients(basePackages = {"org.xcolab.view", "org.xcolab.client" })
 @EnableDiscoveryClient
 public class ColabViewApplication extends SpringBootServletInitializer {
